requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.8
# by the following command:
#
#    pip-compile pyproject.toml
#
annotated-types==0.6.0
    # via pydantic
anyio==4.2.0
    # via httpx
authlib==1.3.0
    # via weaviate-client
blinker==1.7.0
    # via flask
certifi==2024.2.2
    # via
    #   httpcore
    #   httpx
    #   requests
cffi==1.16.0
    # via cryptography
charset-normalizer==3.3.2
    # via requests
click==8.1.7
    # via flask
cryptography==42.0.2
    # via authlib
exceptiongroup==1.2.0
    # via anyio
flask==3.0.1
    # via flask_weaviate (pyproject.toml)
grpcio==1.60.1
    # via
    #   grpcio-health-checking
    #   grpcio-tools
    #   weaviate-client
grpcio-health-checking==1.60.1
    # via weaviate-client
grpcio-tools==1.60.1
    # via weaviate-client
h11==0.14.0
    # via httpcore
httpcore==1.0.2
    # via httpx
httpx==0.26.0
    # via weaviate-client
idna==3.6
    # via
    #   anyio
    #   httpx
    #   requests
importlib-metadata==7.0.1
    # via flask
itsdangerous==2.1.2
    # via flask
jinja2==3.1.3
    # via flask
markupsafe==2.1.5
    # via
    #   jinja2
    #   werkzeug
protobuf==4.25.2
    # via
    #   grpcio-health-checking
    #   grpcio-tools
pycparser==2.21
    # via cffi
pydantic==2.6.0
    # via weaviate-client
pydantic-core==2.16.1
    # via pydantic
requests==2.31.0
    # via weaviate-client
sniffio==1.3.0
    # via
    #   anyio
    #   httpx
typing-extensions==4.9.0
    # via
    #   annotated-types
    #   anyio
    #   pydantic
    #   pydantic-core
urllib3==2.2.0
    # via requests
validators==0.22.0
    # via weaviate-client
weaviate-client==4.4.1
    # via flask_weaviate (pyproject.toml)
werkzeug==3.0.1
    # via flask
zipp==3.17.0
    # via importlib-metadata

# The following packages are considered to be unsafe in a requirements file:
# setuptools
